Реализован Aggregator: вынесены топики Kafka в application.yml, обновлены SensorEventHandler и AggregatorRunner для динамической конфигурации.
