+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/aggregator/AggregatorApplication.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.aggregator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AggregatorApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AggregatorApplication.class, args);
+    }
+}
