+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/aggregator/config/KafkaClient.java
@@ -1,0 +1,4 @@
+package ru.yandex.practicum.aggregator.config;
+
+public interface KafkaClient extends KafkaConsumerClient, KafkaProducerClient {
+}
