+++ telemetry/aggregator/src/main/java/ru/yandex/practicum/aggregator/config/KafkaProducerClient.java
@@ -1,0 +1,10 @@
+package ru.yandex.practicum.aggregator.config;
+
+import org.apache.avro.specific.SpecificRecordBase;
+import org.apache.kafka.clients.producer.KafkaProducer;
+
+public interface KafkaProducerClient {
+    KafkaProducer<String, SpecificRecordBase> getProducer();
+
+    void stopProducer();
+}
