--- telemetry/collector/src/main/java/ru/practicum/kafka/CollectorApplication.java
@@ -1,11 +1,0 @@
-package ru.practicum.kafka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CollectorApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(CollectorApplication.class, args);
-    }
-}
