--- telemetry/collector/src/main/java/ru/practicum/kafka/config/KafkaClient.java
@@ -1,10 +1,0 @@
-package ru.practicum.kafka.config;
-
-import org.apache.avro.specific.SpecificRecordBase;
-import org.apache.kafka.clients.producer.Producer;
-
-public interface KafkaClient {
-    Producer<String, SpecificRecordBase> getProducer();
-
-    void stop();
-}
