--- telemetry/collector/src/main/java/ru/practicum/kafka/service/handler/HubEventHandler.java
@@ -1,9 +1,0 @@
-package ru.practicum.kafka.service.handler;
-
-import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
-
-public interface HubEventHandler {
-    HubEventProto.PayloadCase getMessageType();
-
-    void handle(HubEventProto hubEvent);
-}
