--- telemetry/collector/src/main/java/ru/practicum/kafka/service/handler/SensorEventHandler.java
@@ -1,9 +1,0 @@
-package ru.practicum.kafka.service.handler;
-
-import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
-
-public interface SensorEventHandler {
-    SensorEventProto.PayloadCase getMessageType();
-
-    void handle(SensorEventProto sensorEvent);
-}
