--- telemetry/collector/src/main/java/ru/practicum/kafka/service/handler/hub/DeviceRemovedEventHandler.java
@@ -1,29 +1,0 @@
-package ru.practicum.kafka.service.handler.hub;
-
-import lombok.AccessLevel;
-import lombok.experimental.FieldDefaults;
-import org.springframework.stereotype.Component;
-import ru.practicum.kafka.config.KafkaClient;
-import ru.practicum.kafka.config.KafkaProducerConfig;
-import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
-import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
-
-@Component
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class DeviceRemovedEventHandler extends BaseHubEventHandler<DeviceRemovedEventAvro> {
-    public DeviceRemovedEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
-        super(kafkaClient, kafkaProducerConfig);
-    }
-
-    @Override
-    public DeviceRemovedEventAvro mapToAvro(HubEventProto hubEvent) {
-        return DeviceRemovedEventAvro.newBuilder()
-            .setId(hubEvent.getDeviceRemoved().getId())
-            .build();
-    }
-
-    @Override
-    public HubEventProto.PayloadCase getMessageType() {
-        return HubEventProto.PayloadCase.DEVICE_REMOVED;
-    }
-}
