--- telemetry/collector/src/main/java/ru/practicum/kafka/service/handler/sensor/SwitchSensorEventHandler.java
@@ -1,29 +1,0 @@
-package ru.practicum.kafka.service.handler.sensor;
-
-import lombok.AccessLevel;
-import lombok.experimental.FieldDefaults;
-import org.springframework.stereotype.Component;
-import ru.practicum.kafka.config.KafkaClient;
-import ru.practicum.kafka.config.KafkaProducerConfig;
-import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
-import ru.yandex.practicum.kafka.telemetry.event.SwitchSensorAvro;
-
-@Component
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class SwitchSensorEventHandler extends BaseSensorEventHandler<SwitchSensorAvro> {
-    public SwitchSensorEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
-        super(kafkaClient, kafkaProducerConfig);
-    }
-
-    @Override
-    public SwitchSensorAvro mapToAvro(SensorEventProto sensorEvent) {
-        return SwitchSensorAvro.newBuilder()
-            .setState(sensorEvent.getSwitchSensorEvent().getState())
-            .build();
-    }
-
-    @Override
-    public SensorEventProto.PayloadCase getMessageType() {
-        return SensorEventProto.PayloadCase.SWITCH_SENSOR_EVENT;
-    }
-}
