+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/CollectorApplication.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.collector;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CollectorApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CollectorApplication.class, args);
+    }
+}
