+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/config/KafkaClient.java
@@ -1,0 +1,10 @@
+package ru.yandex.practicum.collector.config;
+
+import org.apache.avro.specific.SpecificRecordBase;
+import org.apache.kafka.clients.producer.Producer;
+
+public interface KafkaClient {
+    Producer<String, SpecificRecordBase> getProducer();
+
+    void stop();
+}
