+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/config/KafkaProducerConfig.java
@@ -1,0 +1,21 @@
+package ru.yandex.practicum.collector.config;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import java.util.Map;
+import java.util.Properties;
+
+@Getter
+@Setter
+@Component
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@ConfigurationProperties(prefix = "collector.kafka.producer")
+public class KafkaProducerConfig {
+    Properties properties;
+    Map<String, String> topics;
+}
