+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/HubEventHandler.java
@@ -1,0 +1,9 @@
+package ru.yandex.practicum.collector.service.handler;
+
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+
+public interface HubEventHandler {
+    HubEventProto.PayloadCase getMessageType();
+
+    void handle(HubEventProto hubEvent);
+}
