+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/SensorEventHandler.java
@@ -1,0 +1,9 @@
+package ru.yandex.practicum.collector.service.handler;
+
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+
+public interface SensorEventHandler {
+    SensorEventProto.PayloadCase getMessageType();
+
+    void handle(SensorEventProto sensorEvent);
+}
