+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/hub/DeviceRemovedEventHandler.java
@@ -1,0 +1,29 @@
+package ru.yandex.practicum.collector.service.handler.hub;
+
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.collector.config.KafkaClient;
+import ru.yandex.practicum.collector.config.KafkaProducerConfig;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
+
+@Component
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class DeviceRemovedEventHandler extends BaseHubEventHandler<DeviceRemovedEventAvro> {
+    public DeviceRemovedEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
+        super(kafkaClient, kafkaProducerConfig);
+    }
+
+    @Override
+    public DeviceRemovedEventAvro mapToAvro(HubEventProto hubEvent) {
+        return DeviceRemovedEventAvro.newBuilder()
+            .setId(hubEvent.getDeviceRemoved().getId())
+            .build();
+    }
+
+    @Override
+    public HubEventProto.PayloadCase getMessageType() {
+        return HubEventProto.PayloadCase.DEVICE_REMOVED;
+    }
+}
