+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/hub/ScenarioRemovedEventHandler.java
@@ -1,0 +1,29 @@
+package ru.yandex.practicum.collector.service.handler.hub;
+
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.collector.config.KafkaClient;
+import ru.yandex.practicum.collector.config.KafkaProducerConfig;
+import ru.yandex.practicum.grpc.telemetry.event.HubEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.ScenarioRemovedEventAvro;
+
+@Component
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class ScenarioRemovedEventHandler extends BaseHubEventHandler<ScenarioRemovedEventAvro> {
+    public ScenarioRemovedEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
+        super(kafkaClient, kafkaProducerConfig);
+    }
+
+    @Override
+    public ScenarioRemovedEventAvro mapToAvro(HubEventProto hubEvent) {
+        return ScenarioRemovedEventAvro.newBuilder()
+            .setName(hubEvent.getScenarioRemoved().getName())
+            .build();
+    }
+
+    @Override
+    public HubEventProto.PayloadCase getMessageType() {
+        return HubEventProto.PayloadCase.SCENARIO_REMOVED;
+    }
+}
