+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/sensor/LightSensorEventHandler.java
@@ -1,0 +1,29 @@
+package ru.yandex.practicum.collector.service.handler.sensor;
+
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.collector.config.KafkaClient;
+import ru.yandex.practicum.collector.config.KafkaProducerConfig;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.LightSensorAvro;
+
+@Component
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class LightSensorEventHandler extends BaseSensorEventHandler<LightSensorAvro> {
+    public LightSensorEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
+        super(kafkaClient, kafkaProducerConfig);
+    }
+
+    public LightSensorAvro mapToAvro(SensorEventProto sensorEvent) {
+        return LightSensorAvro.newBuilder()
+            .setLinkQuality(sensorEvent.getLightSensorEvent().getLinkQuality())
+            .setLuminosity(sensorEvent.getLightSensorEvent().getLuminosity())
+            .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getMessageType() {
+        return SensorEventProto.PayloadCase.LIGHT_SENSOR_EVENT;
+    }
+}
