+++ telemetry/collector/src/main/java/ru/yandex/practicum/collector/service/handler/sensor/SwitchSensorEventHandler.java
@@ -1,0 +1,29 @@
+package ru.yandex.practicum.collector.service.handler.sensor;
+
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.collector.config.KafkaClient;
+import ru.yandex.practicum.collector.config.KafkaProducerConfig;
+import ru.yandex.practicum.grpc.telemetry.event.SensorEventProto;
+import ru.yandex.practicum.kafka.telemetry.event.SwitchSensorAvro;
+
+@Component
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class SwitchSensorEventHandler extends BaseSensorEventHandler<SwitchSensorAvro> {
+    public SwitchSensorEventHandler(KafkaClient kafkaClient, KafkaProducerConfig kafkaProducerConfig) {
+        super(kafkaClient, kafkaProducerConfig);
+    }
+
+    @Override
+    public SwitchSensorAvro mapToAvro(SensorEventProto sensorEvent) {
+        return SwitchSensorAvro.newBuilder()
+            .setState(sensorEvent.getSwitchSensorEvent().getState())
+            .build();
+    }
+
+    @Override
+    public SensorEventProto.PayloadCase getMessageType() {
+        return SensorEventProto.PayloadCase.SWITCH_SENSOR_EVENT;
+    }
+}
