+++ telemetry/pom.xml
@@ -15,6 +15,7 @@
     <modules>
         <module>serialization</module>
         <module>collector</module>
+        <module>aggregator</module>
     </modules>
 
 </project>
