+++ telemetry/serialization/avro-schemas/src/main/avro/telemetry/SensorSnapshot.avdl
@@ -1,0 +1,21 @@
+@namespace("ru.yandex.practicum.kafka.telemetry.event")
+protocol SensorsSnapshotProtocol {
+    import idl "SensorEvent.avdl";
+
+    record SensorsSnapshotAvro {
+        string hub_id;
+        timestamp_ms timestamp;
+        map<SensorStateAvro> sensor_state;
+    }
+
+    record SensorStateAvro {
+        timestamp_ms timestamp;
+        union{
+            ClimateSensorAvro,
+            LightSensorAvro,
+            MotionSensorAvro,
+            SwitchSensorAvro,
+            TemperatureSensorAvro
+        } data;
+    }
+}
