+++ telemetry/serialization/avro-schemas/src/main/java/ru/yandex/practicum/kafka/serializer/BaseAvroDeserializer.java
@@ -1,0 +1,35 @@
+package ru.yandex.practicum.kafka.serializer;
+
+import org.apache.avro.Schema;
+import org.apache.avro.io.BinaryDecoder;
+import org.apache.avro.io.DecoderFactory;
+import org.apache.avro.specific.SpecificDatumReader;
+import org.apache.avro.specific.SpecificRecordBase;
+import org.apache.kafka.common.errors.SerializationException;
+import org.apache.kafka.common.serialization.Deserializer;
+
+import java.io.IOException;
+
+public abstract class BaseAvroDeserializer<T extends SpecificRecordBase> implements Deserializer<T> {
+    final DecoderFactory decoderFactory;
+    final SpecificDatumReader<T> specificDatumReader;
+
+    public BaseAvroDeserializer(Schema schema) {
+        this(DecoderFactory.get(), schema);
+    }
+
+    public BaseAvroDeserializer(DecoderFactory decoderFactory, Schema schema) {
+        this.decoderFactory = decoderFactory;
+        specificDatumReader = new SpecificDatumReader<>(schema);
+    }
+
+    @Override
+    public T deserialize(String topic, byte[] bytes) {
+        BinaryDecoder binaryDecoder = decoderFactory.binaryDecoder(bytes, null);
+        try {
+            return specificDatumReader.read(null, binaryDecoder);
+        } catch (IOException e) {
+            throw new SerializationException("Error deserializing data from topic [" + topic + "]", e);
+        }
+    }
+}
