+++ telemetry/serialization/avro-schemas/src/main/java/ru/yandex/practicum/kafka/serializer/SensorEventDeserializer.java
@@ -1,0 +1,9 @@
+package ru.yandex.practicum.kafka.serializer;
+
+import ru.yandex.practicum.kafka.telemetry.event.SensorEventAvro;
+
+public class SensorEventDeserializer extends BaseAvroDeserializer<SensorEventAvro> {
+    public SensorEventDeserializer() {
+        super(SensorEventAvro.getClassSchema());
+    }
+}
